refactor: moves CustomTooltip component to common folder for reuse
